Add Options and DefaultOptions:
The first and currently only option controls whether string values are assumed to be escaped (using Go escaping syntax). This defaults to true for backwards compatibility. New 'constructors' have been added for type Config for when options are used.

The behaviour (which existed before I introduced Options) where, if a string value contains an invalid escape sequence, the specified default value is returned, seems a bit suspect. It might be better to surface some kind of error to the package consumer so that they are made aware that there is an invalid escape sequence in their INI that needs fixed.

Similarly, the behaviour around keys with no value (e.g. 'foo=' and 'foo[]=') might need to be assessed. Should that yield an error, the empty string, or the specified default value(*)? When the key with no value is part of a string array, should that cause an error, cause an empty string to be part of the yielded slice, or cause the yielded slice to have one fewer elements(*)? The behaviour as of this commit is marked with (*).